+++ hypervisor/src/vm/vm_launch.c
@@ -19,7 +19,7 @@
 
 /* functions */
 
-static errno_t validate_parameters(struct vm *vm, const uint8_t vpc_no, const struct vpc_boot_configuration *boot)
+static errno_t validate_parameters(struct vm *vm, const uint16_t vpc_no, const struct vpc_boot_configuration *boot)
 {
     errno_t ret;
 
@@ -49,7 +49,7 @@ static errno_t validate_parameters(struct vm *vm, const uint8_t vpc_no, const st
  * 本APIではAPIを呼び出した実プロセッサの番号が暗黙の引数として利用される。
  * APIを呼び出した実プロセッサと仮想プロセッサが関連付けされる。
  */
-errno_t vm_launch(struct vm *vm, uint8_t vpc_no, const struct vpc_boot_configuration *boot)
+errno_t vm_launch(struct vm *vm, uint16_t vpc_no, const struct vpc_boot_configuration *boot)
 {
     errno_t ret;
     struct vpc *vpc;
