+++ hypervisor/src/vm/vm_proc_no.c
@@ -17,12 +17,12 @@
 
 /* functions */
 
-uint8_t vm_virtual_proc_no(struct vm *vm, uint8_t physical_no)
+uint16_t vm_virtual_proc_no(struct vm *vm, uint16_t physical_no)
 {
     return vm->proc_map.virtual[physical_no];
 }
 
-uint8_t vm_physical_proc_no(struct vm *vm, uint8_t virtual_no)
+uint16_t vm_physical_proc_no(struct vm *vm, uint16_t virtual_no)
 {
     return vm->proc_map.physical[virtual_no];
 }
